Fixes on DFT-related code

+++ src/Dechorder/Base.hs
@@ -4,13 +4,13 @@ import           Data.Vector
 import           Data.Complex
 
 type SampleRate = Int
-type Duration = Float
+type Duration = Double
 
-type Frequency = Float
-type Amplitude = Float
+type Frequency = Double
+type Amplitude = Double
 
-type SampleChunk = Vector Float
-type SampleChunkF = Vector (Complex Float)
+type SampleChunk = Vector Double
+type SampleChunkF = Vector (Complex Double)
 type MagnitudeChunk = Vector Amplitude
 
 data Key = C | Db | D | Eb | E | F | Gb | G | Ab | A | Bb | B
